(refactor) file structure
