+++ packages/esm-billing-app/src/claims/dashboard/claims-breakdown/claims-breakdown.scss
