+++ packages/esm-billing-app/src/claims/dashboard/claims-header/claims-header.scss
