+++ packages/esm-billing-app/src/claims/dashboard/claims-table.component.tsx/claims-table.scss
